requirements.txt
pytorch-forecasting==0.9.2
wandb==0.12.7
pandas==1.3.4
matplotlib==3.5.0

requests~=2.26.0
paho-mqtt~=1.6.1